メッセージの修正・Dry-Runの追加

+++ magatama_mw.sh
@@ -83,7 +83,7 @@ case ${1} in
             if [ "$?" -ne "0" ]; then
                 # Notify Error.
                 MSG="${CFG_MW_MSG}"
-                MSG_STATUS="Post-update script error."
+                MSG_STATUS="${CFG_MW_MSG_ERR_POST}"
                 if [ -f ./magatama_notify.sh ]; then
                     . ./magatama_notify.sh
                 fi
@@ -102,8 +102,15 @@ case ${1} in
     maintenance)
         # ForceFlag - Force Run Update
         if [ "$2" = "force" ]; then
+            FLG_DRY=0;
             FLG_FORCE=1;
             MSG_STATUS=${CFG_MW_MSG_FORCE}
+        fi
+
+        if [ "$2" = "dry-run" ]; then
+            FLG_DRY=1;
+            FLG_FORCE=0;
+            MSG_STATUS=${CFG_MW_MSG_DRYRUN}
         fi
 
         if [ "${FLG_FORCE}" != "1" ]; then
@@ -113,30 +120,45 @@ case ${1} in
         # if retuen code eq 100, have yum-repos updates.
         if [ "$?" -eq "100" -o "${FLG_FORCE}" = "1" ]; then
 
-            # Run Pre update script
-            if [ -f "/root/bin/magatama_mw_pre.sh" ]; then
-                /root/bin/magatama_mw_pre.sh
-            fi
-
-            # Error Handring
-            if [ "$?" -ne "0" ]; then
-                # Notify Error.
+            if [ "${FLG_DRY}" = "1" ]; then
+                # ---
+                # is Dry-Run
                 MSG="${CFG_MW_MSG}"
-                MSG_STATUS="Pre-update script error."
+                MSG_STATUS="${CFG_MW_MSG_DRYRUN}"
                 if [ -f ./magatama_notify.sh ]; then
-                    # put MSG and MSG_STATUS
                     . ./magatama_notify.sh
                 fi
-                exit 1
-            fi
-            
-            # Update KUSANAGI
-            yum update -y -q > /dev/null
-            if [ "$?" -eq "0" ]; then
-                # Make Maintenance flag and write status message.
-                touch "${FLG_MW_FN}" && echo "${MSG_STATUS}" > ${FLG_MW_FN}
-                # reboot
-                sync && sync && sync && shutdown -r now && exit 0
+                exit 0
+
+            else
+                # ---
+                # is Update
+
+                # Run Pre update script
+                if [ -f "/root/bin/magatama_mw_pre.sh" ]; then
+                    /root/bin/magatama_mw_pre.sh
+                fi
+
+                # Error Handring
+                if [ "$?" -ne "0" ]; then
+                    # Notify Error.
+                    MSG="${CFG_MW_MSG}"
+                    MSG_STATUS="${CFG_MW_MSG_ERR_PRE}"
+                    if [ -f ./magatama_notify.sh ]; then
+                        # put MSG and MSG_STATUS
+                        . ./magatama_notify.sh
+                    fi
+                    exit 1
+                fi
+                
+                # Update KUSANAGI
+                yum update -y -q > /dev/null
+                if [ "$?" -eq "0" ]; then
+                    # Make Maintenance flag and write status message.
+                    touch "${FLG_MW_FN}" && echo "${MSG_STATUS}" > ${FLG_MW_FN}
+                    # reboot
+                    sync && sync && sync && shutdown -r now && exit 0
+                fi
             fi
         fi
     ;;
